### 2018-01-29 #556 @whedon
[REVIEW]: CheckQC: Quick quality control of Illumina sequencing runs
**Submitting author:** @johandahlberg (<a href="http://orcid.org/0000-0001-6962-1460">Johan Dahlberg</a>)
**Repository:** <a href="https://github.com/Molmed/checkQC" target ="_blank">https://github.com/Molmed/checkQC</a>
**Version:** v1.2.0
