### 2018-01-29 #556 @pjotrp
Dear @johandahlberg, thanks for your interesting submission. Do you mind going through above check list and make sure the reviewer can check the boxes ahead of his going through it? Thanks.
