### 2018-01-29 #556 @johandahlberg
Hi @pjotrp!

Thank you for taking this under consideration for publication. I see that @brainstorm has been assigned to review this paper. While I believe that Roman is very well qualified to review this submission, I believe that there might be a conflict of interest, since we are currently collaborating on another project.

As for the checkboxes do I understand you correctly that you want me to look at them and make sure that they can be checked by the reviewer? Not that I should actually check them?
