+++ README.md
@@ -24,11 +24,9 @@ Instrument types supported in checkQC are the following:
 
 Install instructions
 --------------------
-CheckQC **requires Python 3.5** (or higher to run). Furthermore, right now the Illumina Interop
-library needs to be installed separately before moving on to installing checkqc.
+CheckQC **requires Python 3.5** (or higher to run). CheckQC can be installed with pip.
 
 ```
-pip install -f https://github.com/Illumina/interop/releases/tag/v1.1.1 interop
 pip install checkqc
 ```
 
