### 2018-01-29 #556 @pjotrp
@brainstorm, what is the shared project you are working on? Is it overlapping in any way with this submission? Are you confident you can review independently?

@johandahlberg correct. You can't check them yourself. It is just to prevent some unnecessary work. Just state you are confident they can be checked.
