### 2018-01-29 #556 @johandahlberg
@pjotrp I believe we should be able to tick the boxes, with the exception that all the references do not have DOIs. Since some of the software references are websites, since to the best of my knowledge there are no other references for them. Is that alright?
