### 2018-01-30 #556 @brainstorm
@pjotrp Arteria/StackStorm (projects I'm collaborating with right now) **could** use CheckQC in some way but they don't by default. Similar analogy: would reviewing a particular tool enter in a conflict of interest if I was using/developing, say, Galaxy core?

At least that's my interpretation, hope I'm not missing the point, this is important to clarify early on.
