### 2018-01-30 #556 @johandahlberg
In the interest of full disclosure I think that it should be noted that me and @brainstorm are co-authoring a paper about the Arteria project. Unfortunately I interpret this as grounds for a conflict of interest.

I would like to point out that I still believe that @brainstorm is very well qualified to review this submission. However, to ensure that there can be no ethical ambiguities I think that it would be better if another reviewer was assigned. I will of course defer to any decision made by the editor(s).
