### 2018-01-30 #556 @pjotrp
Lets ask the EIC. Dear @arfon, I think with Roman @brainstorm that he is not having a COI because he happens to be on a different paper with the author. Roman came recommended by another co-worker and sees no conflict of interest himself. I have the intention of giving an all clear. Are you OK with this?
