### 2018-01-31 #556 @arfon
> Are you OK with this?

👍 I think this is OK. Thanks for disclosing this @johandahlberg (and for checking @pjotrp).
