### 2018-02-01 #556 @brainstorm
Thanks @pjotrp, going ahead ;)

@johandahlberg, first bump in the road:

```
$ checkqc 171128_A00119_0035_BH2CTLDSXX
INFO     ------------------------
INFO     Starting checkQC (1.2.0)
INFO     ------------------------
INFO     Runfolder is: 171128_A00119_0035_BH2CTLDSXX/
INFO     No config file specified, using default config from /home/limsadmin/.miniconda3/lib/python3.6/site-packages/checkQC/default_config/config.yaml.
INFO     Run summary
INFO     -----------
INFO     Instrument and reagent version: novaseq_v1
INFO     Read length: 150-150
INFO     Enabled handlers and their config values were:
INFO     	ClusterPFHandler Error=unknown Warning=1400
INFO     	Q30Handler Error=unknown Warning=75
INFO     	ErrorRateHandler Error=unknown Warning=5
INFO     	ReadsPerSampleHandler Error=700 Warning=unknown
INFO     	UndeterminedPercentageHandler Error=10 Warning=unknown
ERROR    Could not identify a Stats.json file at: 171128_A00119_0035_BH2CTLDSXX/Unaligned/Stats/Stats.json. This file is created by bcl2fastq, please ensure that you have run bcl2fastq on this runfolder before running checkqc.
ERROR    Finished with fatal qc errors and will exit with non-zero exit status.

$ find 171128_A00119_0035_BH2CTLDSXX -iname Stats.json
171128_A00119_0035_BH2CTLDSXX/Stats/Stats.json
```
