### 2018-02-01 #556 @brainstorm
Also, regarding the statement of need and towards publishing your paper, could you comment/compare with [basecallQC](https://bioconductor.org/packages/release/bioc/vignettes/basecallQC/inst/doc/basecallQC.html)?
