### 2018-02-01 4970abc
Introducing the parser config concept
In order to be able to configure the parsers I've introduced the concept
of a parser config, which is basically just a dict containing the
configuration of the different parsers under their respective class
names. This dict is then passed to the parsers when they are
instantiated.

I have then used to to be able to configure where the output of
bcl2fastq has been written to, which can then be used to located the
Stats.json file.

This closes #47
