+++ checkQC/parsers/stats_json_parser.py
@@ -4,7 +4,7 @@ import os
 import logging
 
 from checkQC.parsers.parser import Parser
-from checkQC.exceptions import StatsJsonNotFound
+from checkQC.exceptions import StatsJsonNotFound, ConfigurationError
 
 log = logging.getLogger(__name__)
 
@@ -21,18 +21,30 @@ class StatsJsonParser(Parser):
     The subscribers decide which of these values they are iterested in.
     """
 
-    def __init__(self, runfolder, *args, **kwargs):
+    def __init__(self, runfolder, parser_configurations, *args, **kwargs):
         """
         Create a StatsJsonParser instance for the specified runfolder
-
         :param runfolder: path to the runfolder to parse
+        :param parser_configurations: dict containing any extra configuration required by
+        the parser under class name key
         """
         super().__init__(*args, **kwargs)
-        self.file_path = os.path.join(runfolder, "Unaligned", "Stats", "Stats.json")
+
+        self.parser_conf = parser_configurations[self.__class__.__name__]
+        bcl2fastq_output_path = self.parser_conf.get("bcl2fastq_output_path")
+        if not bcl2fastq_output_path:
+            raise ConfigurationError("The configuration must contain the key bcl2fastq_output_path, specifying "
+                                     "where the bcl2fastq output is, relative to the runfolder root.")
+
+        self.file_path = os.path.join(runfolder, bcl2fastq_output_path, "Stats", "Stats.json")
         if not os.path.exists(self.file_path):
             log.error("Could not identify a Stats.json file at: {}. This file is "
                       "created by bcl2fastq, please ensure that you have run "
-                      "bcl2fastq on this runfolder before running checkqc.".format(self.file_path))
+                      "bcl2fastq on this runfolder before running checkqc."
+                      "If this file is not located under <RUNFOLDER>/Data/Intensities/BaseCalls/Stats/Stats.json "
+                      "which is the default option for bcl2fastq, you can specify where the 'Stats' directory is "
+                      "located by changing the 'bcl2fastq_output_path' in the 'StatsJsonParser' part of the "
+                      "checkqc configuration file.".format(self.file_path))
             raise StatsJsonNotFound("Could not find a Stats.json file at: {}".format(self.file_path))
 
     def run(self):
