+++ tests/parsers/test_stats_json_parser.py
@@ -26,7 +26,8 @@ class TestStatsJsonParser(unittest.TestCase):
 
     runfolder = os.path.join(os.path.dirname(__file__), "..", "resources",
                              "170726_D00118_0303_BCB1TVANXX")
-    stats_json_parser = StatsJsonParser(runfolder=runfolder)
+    parser_configs = {"StatsJsonParser": {"bcl2fastq_output_path": "Data/Intensities/BaseCalls"}}
+    stats_json_parser = StatsJsonParser(runfolder=runfolder, parser_configurations=parser_configs)
     subscriber = Receiver()
     stats_json_parser.add_subscribers(subscriber)
     stats_json_parser.run()
@@ -37,8 +38,12 @@ class TestStatsJsonParser(unittest.TestCase):
     def test_init_stats_json_parser_without_stats_json(self):
         with self.assertLogs() as cm:
             with self.assertRaises(StatsJsonNotFound):
-                StatsJsonParser("")
-        expected_log = 'ERROR:checkQC.parsers.stats_json_parser:Could not identify a Stats.json file at: ' \
-                       'Unaligned/Stats/Stats.json. This file is created by bcl2fastq, please ensure that you ' \
-                       'have run bcl2fastq on this runfolder before running checkqc.'
+                StatsJsonParser("", parser_configurations=self.parser_configs)
+        expected_log = "ERROR:checkQC.parsers.stats_json_parser:Could not identify a Stats.json file at: " \
+                       "Data/Intensities/BaseCalls/Stats/Stats.json. This file is created by bcl2fastq, please ensure that you " \
+                       "have run bcl2fastq on this runfolder before running checkqc." \
+                       "If this file is not located under <RUNFOLDER>/Data/Intensities/BaseCalls/Stats/Stats.json " \
+                       "which is the default option for bcl2fastq, you can specify where the 'Stats' directory is " \
+                       "located by changing the 'bcl2fastq_output_path' in the 'StatsJsonParser' part of the " \
+                       "checkqc configuration file."
         self.assertIn(expected_log, cm.output)
