### 2018-02-01 9a431c5
Error message from StatJsonParser config problems
