### 2018-02-01 d0cf304
Improving tests for StatsJsonParser and config
