### 2018-02-01 #556 @johandahlberg
I have addressed the problems you were having with not finding the `Stats.json` file. The path to the `Stats` directory is configurable now, and the default should match up with the default settings for `bcl2fastq`. So far this has only been merged into master. Before moving on to creating a new release I want to know if that means that I should also update the release version in `codemeta.json`? And if you want me to create a release now for you to test, or if you want to keep on testing straight from the master branch? I'm unsure what your prefered process is with regards to that.

Thanks for bringing basecallQC to my attention. I had not seen it before. From what I understand based on the documentation you referenced, the purpose of the basecallQC library is somewhat different from what we are trying to achive with CheckQC.

The features listed for basecallQC include:
 - cleaning samplesheets
 - convering samplesheets between bcl2fastq version
 - creating base masks to be used by bcl2fastq
 - creating summary tables and plots based on demultiplexed data

On the first three points, there is no overlap in functionality. On the last point there might be a small amount of overlap in that both programs/libraries read data from the Illumina runfolder, but with the difference that CheckQC is expliclity constructed to be used to automatically check the values gathered against a set of criteria. I therefore see these to as complementary to each other. Would you like me to add a reference and a short section comparing CheckQC with basecallQC to the paper?
