### 2018-02-02 #556 @brainstorm
Thanks for the fix Johan!

I think that the editor will link/refer to the latest version and assign a DOI, so yes, in the interest of having a working package "out of the box" as reflected in the README, you should update accordingly.

Regarding basecallQC, I don't think a whole section is worth it, but a small mention as you just did here should suffice.

I'm done here, nice tool overall, when you are done with those latest additions, please at-ping Pjotr so that he can generate the DOI and do the rest ;)
