### 2018-02-02 55f1942
Displayed threshold is now divided by the number of samples in the lane
