### 2018-02-02 7d359b8
error and warning thresholds as variables
