+++ checkQC/handlers/reads_per_sample_handler.py
@@ -39,20 +39,25 @@ class ReadsPerSampleHandler(QCHandler):
                 sample_id = sample_id_info["SampleId"]
                 sample_total_reads = sample_id_info["NumberReads"] / pow(10, 6)
 
-                if self.error() != self.UNKNOWN and sample_total_reads < (float(self.error()) / float(nbr_of_samples)):
+                if self.error() != self.UNKNOWN:
+                    error_threshold = float(self.error()) / float(nbr_of_samples)
+                if self.warning() != self.UNKNOWN:
+                    warning_threshold = float(self.warning()) / float(nbr_of_samples)
+
+                if self.error() != self.UNKNOWN and sample_total_reads < error_threshold:
                     yield QCErrorFatal("Number of reads for sample {} was too low on lane {}, "
                                        "it was: {:.3f} M".format(sample_id, lane_nbr, sample_total_reads),
                                        ordering=lane_nbr,
                                        data={"lane": lane_nbr, "number_of_samples": nbr_of_samples,
                                              "sample_id": sample_id, "sample_reads": sample_total_reads,
-                                             "threshold": float(self.error()) / float(nbr_of_samples)})
+                                             "threshold": error_threshold})
                 elif self.warning() != self.UNKNOWN and \
-                                sample_total_reads < (float(self.warning()) / float(nbr_of_samples)):
+                                sample_total_reads < warning_threshold:
                     yield QCErrorWarning("Number of reads for sample {} was too low on lane {}, "
                                          "it was: {:.3f} M".format(sample_id, lane_nbr, sample_total_reads),
                                          ordering=lane_nbr,
                                          data={"lane": lane_nbr, "number_of_samples": nbr_of_samples,
                                                "sample_id": sample_id, "sample_reads": sample_total_reads,
-                                               "threshold": float(self.warning()) / float(nbr_of_samples)})
+                                               "threshold": warning_threshold})
                 else:
                     continue
