### 2018-02-05 fb15074
Add reference to basecallQC to paper
