### 2018-02-05 #556 @johandahlberg
@brainstorm I've fixed your suggestions now. And I've created a new release: https://github.com/Molmed/checkQC/releases/tag/v1.3.0 This is now available on PyPI.

@pjotrp If I understand @brainstorm above this is done now. As noted above, there is a new release (v1.3.0), and I've updated the `codemeta.json` to reflect this.
