### 2018-02-05 #556 @johandahlberg
@arfon Thank you! I have created a DOI for the specific version (v1.3.0) here: https://doi.org/10.5281/zenodo.1166052. And the DOI for any version of the software is this one: https://doi.org/10.5281/zenodo.1166051 I'm unsure about which one you need, so now you have both.
